requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
boto3
python-dotenv
apscheduler
pydantic
yandex_cloud_ml_sdk
python-multipart 
